Minor naming refactoring to reflect the codebase and logic
