+++ src/domain/project/project.resolver.ts
@@ -1,6 +1,6 @@
 import { UseGuards } from '@nestjs/common';
 import { Args, Mutation, Query, Resolver } from '@nestjs/graphql';
-import { GqlAuthGuard } from '@utils/authentication/graphql.guard';
+import { GqlAuthGuard } from '@utils/auth/graphql.guard';
 import { Roles } from '@utils/decorators/roles.decorator';
 import { Profiling } from '@utils/logging/logging.profiling.decorator';
 import { AspectInput } from '@domain/aspect/aspect.dto';
