+++ src/domain/tagset/tagset.resolver.ts
@@ -1,7 +1,7 @@
 import { UseGuards } from '@nestjs/common';
 import { Args, Mutation, Resolver } from '@nestjs/graphql';
 import { Roles } from '@utils/decorators/roles.decorator';
-import { GqlAuthGuard } from '@utils/authentication/graphql.guard';
+import { GqlAuthGuard } from '@utils/auth/graphql.guard';
 import { RestrictedGroupNames } from '@domain/user-group/user-group.entity';
 import { Tagset } from './tagset.entity';
 import { TagsetService } from './tagset.service';
