+++ src/domain/user-group/user-group.resolver.fields.ts
@@ -2,7 +2,7 @@ import { UseGuards } from '@nestjs/common';
 import { Resolver } from '@nestjs/graphql';
 import { Parent, ResolveField } from '@nestjs/graphql';
 import { Roles } from '@utils/decorators/roles.decorator';
-import { GqlAuthGuard } from '@utils/authentication/graphql.guard';
+import { GqlAuthGuard } from '@utils/auth/graphql.guard';
 import { RestrictedGroupNames, UserGroup } from './user-group.entity';
 import { UserGroupService } from './user-group.service';
 import { UserGroupParent } from './user-group-parent.dto';
