+++ src/utils/auth/auth.module.ts
@@ -1,13 +1,13 @@
 import { forwardRef, Module } from '@nestjs/common';
 import { PassportModule } from '@nestjs/passport';
 import { UserModule } from '@domain/user/user.module';
-import { OidcStrategy } from './oidc.strategy';
+import { OidcBearerStrategy } from './oidc.bearer.strategy';
 import { AuthService } from './auth.service';
 @Module({
   imports: [
     PassportModule.register({ session: false, defaultStrategy: 'bearer' }),
     forwardRef(() => UserModule),
   ],
-  providers: [OidcStrategy, AuthService],
+  providers: [OidcBearerStrategy, AuthService],
 })
-export class AuthenticationModule {}
+export class AuthModule {}
